Boot into x86-64 long mode
The kernel is now built for x86-64 and puts the CPU into long mode during
boot, when the CPU supports it.

Multiboot 1 can only hand over a 32-bit protected mode machine with paging
off, and long mode cannot be entered from there directly: it needs paging
already enabled, and long mode paging needs a four-level table. So boot.S
stays 32-bit long enough to check CPUID for the long mode bit, identity
map the first 1 GiB with 2 MiB pages, set PAE and EFER.LME and turn on
paging, then far-jump through a 64-bit code segment.

A CPU without long mode gets a legible message on the screen and on COM1
rather than a triple fault. There is deliberately no 32-bit fallback: the
rest of the kernel is compiled for x86-64, so there would be nothing to
fall back into. One boot image cannot be both.

What the port touched:

- Four-level paging. Kernel and user share PML4 entry 0, so a process has
  its own PML4 and PDPT with the kernel's page directory installed in the
  first slot. The user bit is set at PML4 level and clear on the identity
  map, and the CPU ANDs it across all four levels, which is what keeps
  ring 3 out of kernel memory.
- The first 2 MiB huge page is split into 4 KiB pages purely so page zero
  can stay unmapped; a huge page cannot exclude part of itself, and a
  null write that silently lands on physical 0 is much worse than a fault.
- No pusha in long mode, so the stub pushes fifteen registers by hand. In
  exchange the CPU always pushes rsp and ss, so the frame is one shape.
- Syscalls take their arguments in the System V registers now.
- ELF64 user binaries; argv is an array of 64-bit pointers, which the
  spawn path had to be taught before it read half of each one.
- 64-bit division is a native instruction here, so kprintf and the user
  printf can finally print 64-bit quantities without libgcc.
- -mno-red-zone is now load-bearing: without it an interrupt lands on the
  128 bytes below rsp that a leaf function assumes are its own.

One bug worth naming: the GDT puts user data at index 3 and user code at
index 4, but the selector constants had been written for indices 4 and 5.
That made CS name the TSS descriptor, and the iretq into ring 3 faulted
with the TSS selector in the error code.

Verified: all 43 self-tests pass, whell, fish, vim, htop and fastfetch all
run, and the memory map now shows regions above 4 GiB that the 32-bit
kernel could not address.

+++ kernel/include/vfs.h
@@ -58,6 +58,6 @@ int vfs_getcwd(struct process *p, char *buf, uint32_t size);
 /* Read a whole file into a freshly kmalloc'd buffer.  The caller owns the
  * buffer and must kfree it.  Used by the program loader. */
 int vfs_read_file(struct process *p, const char *path, void **data_out,
-                  uint32_t *size_out);
+                  uint64_t *size_out);
 
 #endif /* WOS_VFS_H */
